+++ lib/dissociated_introspection/block.rb
@@ -1,0 +1,11 @@
+module DissociatedIntrospection
+  Block = Struct.new(:ast, :type)
+
+  def Block.source
+    :source
+  end
+
+  def Block.to_h
+    super.merge(source: source)
+  end
+end