+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.8.0 - 2017-03-24
+### Enhancement
+- DissociatedIntrospection::RubyClass#class_defs returns same api as #defs
+- DissociatedIntrospection::RubyClass#class_method_calls inspect method name and its arguments ie. attr_reader :name
+
+### Deprecated
+ - DissociatedIntrospection::RubyClass#is_class? changed to #class?
+ - DissociatedIntrospection::RubyClass#has_parent_class? changed to #parent_class?
+ - DissociatedIntrospection::RubyCode#to_ruby_str changed to #source
+
 ## 0.7.1 - 2016-10-20
 ### Fix
 - `DissociatedIntrospection::RubyClass#defs` fixed parse issue when class has one method.